refactor: js2ts graph
