+++ src/graphlib/index.ts
@@ -1,8 +1,7 @@
-import Graph = require('./graph');
+export { Graph } from './graph';
 import { isAcyclic } from './alg/is-acyclic';
 import { postorder } from './alg/postorder';
 
-export { Graph };
 export const alg = {
   isAcyclic,
   postorder,
